+   Converted networking to use netty 4 library 	-   This came with addition of RSA & ISAAC encryption/decryption +   Changed client src to a better 666 src +  Added 'lobby server' to be the main connection server 	-   'World server' support is easy and adding another world is as simple as pressing run 	-   Each world no longer has their own lobby and we have one big lobby server +   This might seem like small tasks but check the second photo for context as to how long it takes to do these things right 	-   Credits to Walied for answers & support

+++ src/org/redrune/game/GameFlags.java
@@ -16,6 +16,6 @@ public class GameFlags {
 	/**
 	 * The id of the world that will be ran
 	 */
-	public static int worldId;
+	public static byte worldId;
 	
 }
